+++ src/shared/copy-arc.js
@@ -23,16 +23,16 @@ module.exports = function copyArc (params, paths, callback) {
     let done = `Hydrated app with Architect manifest`
     let start = update.start(`Hydrating app with Architect manifest`)
 
-    function _done (err) {
-      let cmd = 'copy'
-      if (err) print({ cmd, err, start, update }, callback)
-      else print({ cmd, start, done, update }, callback)
-    }
-    series(paths.map(dest => {
+    let destinations = Object.entries(paths).map(p => p[1])
+    series(destinations.map(dest => {
       return function copier (callback) {
         copy(join(dest, 'shared'), params, callback)
       }
-    }), _done)
+    }), function _done (err) {
+      let cmd = 'copy'
+      if (err) print({ cmd, err, start, update }, callback)
+      else print({ cmd, start, done, update }, callback)
+    })
   }
   else callback()
 }